Implement hash dir and feh asks
Now takes 2 arguments, a hash directory and an import directory

It uses the file names in the hash directory as known hashes, and
imports files in the import directory one by one, comparing it to the
known hashes and opening a feh window to delete a potential duplicate
(press enter to delete). Imports it if it wasn't deleted.